backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-dotenv
passlib[bcrypt]
PyJWT
bcrypt